+++ RELEASE-NOTES.md
@@ -1,18 +1,28 @@
 ddth-cache-adapter release notes
 ================================
 
+0.3.0 - 2015-11-01
+------------------
+
+- Custom cache entry de/serializations.
+- 2 built-in cache entry serializers: `DefaultCacheEntrySerializer` and `KryoCacheEntrySerializer`.
+
+
 0.2.1.8 - 2015-06-30
 --------------------
+
 - Bugs fixed: Redis' compact mode is not set.
 
 
 0.2.1.7 - 2014-11-24
 --------------------
+
 - Bugs fixed: cache entry's TTL.
 
 
 0.2.0.2 - 2014-10-13
 --------------------
+
 - New caches: `ThreadLocalCache` and `LocalRemoteCache`.
 - Some improvements.
 - Bugs fixed.
@@ -20,14 +30,17 @@ ddth-cache-adapter release notes
 
 0.1.2 - 2014-07-30
 ------------------
+
 - Fix a StackOverflowError bug in `GuavaCache` class.
 
 
 0.1.1 - 2014-07-28
 ------------------
+
 - Redis cache: Support compact and non-compact mode (default mode: non-compact).
 
 
 0.1.0 - 2014-03-24
 ------------------
+
 - First release.